--- a/model/metric/processor.go
+++ b/model/metric/processor.go
@@ -15,20 +15,4 @@
 // specific language governing permissions and limitations
 // under the License.
 
-package transaction
-
-import (
-	"testing"
-
-	"github.com/stretchr/testify/assert"
-
-	pr "github.com/elastic/apm-server/processor"
-)
-
-func TestImplementProcessorInterface(t *testing.T) {
-	p := NewProcessor()
-	assert.NotNil(t, p)
-	_, ok := p.(pr.Processor)
-	assert.True(t, ok)
-	assert.IsType(t, &processor{}, p)
-}
+package metric