… model/error/_meta/fields.yml
… model/error/event_test.go
… model/error/generated/schema/payload.go
… model/error/payload_test.go
… model/metric/_meta/fields.yml
… model/metric/counter.go
… model/metric/event.go
… model/metric/gauge.go
… model/metric/generated/schema/payload.go
… model/metric/payload_test.go
… model/metric/summary.go
… model/sourcemap/_meta/fields.yml
… model/sourcemap/generated/schema/payload.go
… model/sourcemap/payload_test.go
… model/transaction/_meta/fields.yml
… model/transaction/event.go
… model/transaction/event_test.go
… model/transaction/generated/schema/payload.go
… model/transaction/payload_test.go
… model/transaction/span.go
… model/transaction/span_context.go
… model/transaction/span_context_test.go
… model/transaction/span_test.go
… processor/error/package_tests/processor_test.go
… processor/error/payload.go
… processor/error/processor.go
… processor/metric/metric.go
… processor/metric/package_tests/json_schema_test.go
… processor/metric/package_tests/processor_test.go
… processor/metric/processor.go