This moves error, transaction, sourcemap and metric models into directories in the "model" directory. It changes processors to be simple structs which can potentially override Decode/Validate (only sourcemaps overrides these). This removes a bunch of boilerplate from the processors. It will help move towards decoding and validating models individually, which is needed in the future for intake-v2.

I also got rid of the healthcheck processor, which is now a simple http.Handler. 

There are still more we can do to separate models from processing, for example, counting (validation, transformations etc.) is spread between models and processor now, but I was unable to completely fix it in this PR. Subsequent PRs will address this.