@simitt 2018-07-17
@roncohen could you please also provide benchmark results? I don't expect any big changes, but we should rely on numbers for refactoring PRs.