@roncohen 2018-07-17
benchmark results show there that are no difference in the number of allocations at all. The performance number for this PR looks better than `master`, but I suspect there's no real difference and that any difference in the benchmarks results stem from the computer idling a little more or a little less at any given time. 

<details>
<summary>Benchmark details</summary>


```
benchmark                                                                              old ns/op     new ns/op     delta
BenchmarkBackendProcessor/TestProcessErrorMinimalServiceValidate-4                     8422          8101          -3.81%
BenchmarkBackendProcessor/TestProcessErrorMinimalServiceDecode-4                       3788          3793          +0.13%
BenchmarkBackendProcessor/TestProcessErrorMinimalServiceTransform-4                    7361          7340          -0.29%
BenchmarkBackendProcessor/TestProcessErrorMinimalServiceProcessRequest-4               16318         16378         +0.37%
BenchmarkBackendProcessor/TestProcessErrorMinimalProcessValidate-4                     10393         10448         +0.53%
BenchmarkBackendProcessor/TestProcessErrorMinimalProcessDecode-4                       4277          4423          +3.41%
BenchmarkBackendProcessor/TestProcessErrorMinimalProcessTransform-4                    7747          7936          +2.44%
BenchmarkBackendProcessor/TestProcessErrorMinimalProcessProcessRequest-4               19483         20542         +5.44%
BenchmarkBackendProcessor/TestProcessErrorFullValidate-4                               62240         62598         +0.58%
BenchmarkBackendProcessor/TestProcessErrorFullDecode-4                                 15654         16423         +4.91%
BenchmarkBackendProcessor/TestProcessErrorFullTransform-4                              39192         39852         +1.68%
BenchmarkBackendProcessor/TestProcessErrorFullProcessRequest-4                         112946        114968        +1.79%
BenchmarkBackendProcessor/TestProcessErrorNullValuesValidate-4                         37541         38295         +2.01%
BenchmarkBackendProcessor/TestProcessErrorNullValuesDecode-4                           8254          7937          -3.84%
BenchmarkBackendProcessor/TestProcessErrorNullValuesTransform-4                        19331         19176         -0.80%
BenchmarkBackendProcessor/TestProcessErrorNullValuesProcessRequest-4                   64054         63170         -1.38%
BenchmarkBackendProcessor/TestProcessErrorAugmentedIPValidate-4                        8951          8617          -3.73%
BenchmarkBackendProcessor/TestProcessErrorAugmentedIPDecode-4                          5853          4164          -28.86%
BenchmarkBackendProcessor/TestProcessErrorAugmentedIPTransform-4                       8616          7964          -7.57%
BenchmarkBackendProcessor/TestProcessErrorAugmentedIPProcessRequest-4                  22934         18178         -20.74%
BenchmarkBackendProcessor/TestProcessErrorMinimalPayloadExceptionValidate-4            5372          3749          -30.21%
BenchmarkBackendProcessor/TestProcessErrorMinimalPayloadExceptionDecode-4              2966          2788          -6.00%
BenchmarkBackendProcessor/TestProcessErrorMinimalPayloadExceptionTransform-4           7079          6376          -9.93%
BenchmarkBackendProcessor/TestProcessErrorMinimalPayloadExceptionProcessRequest-4      12853         11789         -8.28%
BenchmarkBackendProcessor/TestProcessErrorMinimalPayloadLogValidate-4                  7699          5798          -24.69%
BenchmarkBackendProcessor/TestProcessErrorMinimalPayloadLogDecode-4                    3246          2696          -16.94%
BenchmarkBackendProcessor/TestProcessErrorMinimalPayloadLogTransform-4                 7851          6450          -17.84%
BenchmarkBackendProcessor/TestProcessErrorMinimalPayloadLogProcessRequest-4            19496         13401         -31.26%
BenchmarkFrontendProcessor/TestProcessErrorFrontendValidate-4                          28232         22405         -20.64%
BenchmarkFrontendProcessor/TestProcessErrorFrontendDecode-4                            9989          8926          -10.64%
BenchmarkFrontendProcessor/TestProcessErrorFrontendTransform-4                         832673        690101        -17.12%
BenchmarkFrontendProcessor/TestProcessErrorFrontendProcessRequest-4                    1029582       731704        -28.93%
BenchmarkFrontendProcessor/TestProcessErrorFrontendNoSmapValidate-4                    25161         20880         -17.01%
BenchmarkFrontendProcessor/TestProcessErrorFrontendNoSmapDecode-4                      7749          7411          -4.36%
BenchmarkFrontendProcessor/TestProcessErrorFrontendNoSmapTransform-4                   555828        539060        -3.02%
BenchmarkFrontendProcessor/TestProcessErrorFrontendNoSmapProcessRequest-4              591717        566949        -4.19%
BenchmarkFrontendProcessor/TestProcessErrorFrontendMinifiedSmapValidate-4              32354         31215         -3.52%
BenchmarkFrontendProcessor/TestProcessErrorFrontendMinifiedSmapDecode-4                10252         8695          -15.19%
BenchmarkFrontendProcessor/TestProcessErrorFrontendMinifiedSmapTransform-4             75796         70658         -6.78%
BenchmarkFrontendProcessor/TestProcessErrorFrontendMinifiedSmapProcessRequest-4        115295        111418        -3.36%
BenchmarkFrontendProcessor/TestProcessErrorAugmentedUserAgentAndIPValidate-4           8834          8119          -8.09%
BenchmarkFrontendProcessor/TestProcessErrorAugmentedUserAgentAndIPDecode-4             4573          4132          -9.64%
BenchmarkFrontendProcessor/TestProcessErrorAugmentedUserAgentAndIPTransform-4          7961          7033          -11.66%
BenchmarkFrontendProcessor/TestProcessErrorAugmentedUserAgentAndIPProcessRequest-4     17672         16424         -7.06%
BenchmarkProcessor/TestProcessMetricValidate-4                                         80466         77129         -4.15%
BenchmarkProcessor/TestProcessMetricDecode-4                                           10322         9132          -11.53%
BenchmarkProcessor/TestProcessMetricTransform-4                                        15199         13677         -10.01%
BenchmarkProcessor/TestProcessMetricProcessRequest-4                                   102735        101347        -1.35%
BenchmarkProcessor/TestProcessMetricMinimalValidate-4                                  16976         16101         -5.15%
BenchmarkProcessor/TestProcessMetricMinimalDecode-4                                    4754          4077          -14.24%
BenchmarkProcessor/TestProcessMetricMinimalTransform-4                                 4461          3832          -14.10%
BenchmarkProcessor/TestProcessMetricMinimalProcessRequest-4                            22441         21659         -3.48%
BenchmarkProcessor/TestProcessMetricMultipleSamplesValidate-4                          41663         38095         -8.56%
BenchmarkProcessor/TestProcessMetricMultipleSamplesDecode-4                            7998          7451          -6.84%
BenchmarkProcessor/TestProcessMetricMultipleSamplesTransform-4                         23402         19484         -16.74%
BenchmarkProcessor/TestProcessMetricMultipleSamplesProcessRequest-4                    79588         62665         -21.26%
BenchmarkProcessor/TestProcessMetricNullValidate-4                                     27701         24969         -9.86%
BenchmarkProcessor/TestProcessMetricNullDecode-4                                       5418          4855          -10.39%
BenchmarkProcessor/TestProcessMetricNullTransform-4                                    4997          4392          -12.11%
BenchmarkProcessor/TestProcessMetricNullProcessRequest-4                               34942         29590         -15.32%
BenchmarkBackendProcessor/TestProcessTransactionFullValidate-4                         115246        100464        -12.83%
BenchmarkBackendProcessor/TestProcessTransactionFullDecode-4                           19866         17098         -13.93%
BenchmarkBackendProcessor/TestProcessTransactionFullTransform-4                        49920         41893         -16.08%
BenchmarkBackendProcessor/TestProcessTransactionFullProcessRequest-4                   244533        152353        -37.70%
BenchmarkBackendProcessor/TestProcessTransactionNullValuesValidate-4                   71966         53796         -25.25%
BenchmarkBackendProcessor/TestProcessTransactionNullValuesDecode-4                     16439         11750         -28.52%
BenchmarkBackendProcessor/TestProcessTransactionNullValuesTransform-4                  31357         26577         -15.24%
BenchmarkBackendProcessor/TestProcessTransactionNullValuesProcessRequest-4             106171        92811         -12.58%
BenchmarkBackendProcessor/TestProcessSystemNullValidate-4                              11387         10414         -8.54%
BenchmarkBackendProcessor/TestProcessSystemNullDecode-4                                4058          4095          +0.91%
BenchmarkBackendProcessor/TestProcessSystemNullTransform-4                             8718          6188          -29.02%
BenchmarkBackendProcessor/TestProcessSystemNullProcessRequest-4                        20046         16422         -18.08%
BenchmarkBackendProcessor/TestProcessProcessNullValidate-4                             10897         8766          -19.56%
BenchmarkBackendProcessor/TestProcessProcessNullDecode-4                               4160          3774          -9.28%
BenchmarkBackendProcessor/TestProcessProcessNullTransform-4                            6771          6149          -9.19%
BenchmarkBackendProcessor/TestProcessProcessNullProcessRequest-4                       19271         16406         -14.87%
BenchmarkBackendProcessor/TestProcessTransactionMinimalSpanValidate-4                  13598         11517         -15.30%
BenchmarkBackendProcessor/TestProcessTransactionMinimalSpanDecode-4                    5748          4705          -18.15%
BenchmarkBackendProcessor/TestProcessTransactionMinimalSpanTransform-4                 10394         8421          -18.98%
BenchmarkBackendProcessor/TestProcessTransactionMinimalSpanProcessRequest-4            30108         23850         -20.79%
BenchmarkBackendProcessor/TestProcessTransactionMinimalServiceValidate-4               10187         8647          -15.12%
BenchmarkBackendProcessor/TestProcessTransactionMinimalServiceDecode-4                 4810          3756          -21.91%
BenchmarkBackendProcessor/TestProcessTransactionMinimalServiceTransform-4              8311          6604          -20.54%
BenchmarkBackendProcessor/TestProcessTransactionMinimalServiceProcessRequest-4         18834         16943         -10.04%
BenchmarkBackendProcessor/TestProcessTransactionMinimalProcessValidate-4               11452         9958          -13.05%
BenchmarkBackendProcessor/TestProcessTransactionMinimalProcessDecode-4                 4924          4248          -13.73%
BenchmarkBackendProcessor/TestProcessTransactionMinimalProcessTransform-4              8008          7065          -11.78%
BenchmarkBackendProcessor/TestProcessTransactionMinimalProcessProcessRequest-4         21622         18698         -13.52%
BenchmarkBackendProcessor/TestProcessTransactionEmptyValidate-4                        10087         9094          -9.84%
BenchmarkBackendProcessor/TestProcessTransactionEmptyDecode-4                          4136          3995          -3.41%
BenchmarkBackendProcessor/TestProcessTransactionEmptyTransform-4                       7558          6316          -16.43%
BenchmarkBackendProcessor/TestProcessTransactionEmptyProcessRequest-4                  21220         16642         -21.57%
BenchmarkBackendProcessor/TestProcessTransactionAugmentedIPValidate-4                  9470          8481          -10.44%
BenchmarkBackendProcessor/TestProcessTransactionAugmentedIPDecode-4                    4640          3931          -15.28%
BenchmarkBackendProcessor/TestProcessTransactionAugmentedIPTransform-4                 7726          6417          -16.94%
BenchmarkBackendProcessor/TestProcessTransactionAugmentedIPProcessRequest-4            18982         17574         -7.42%
BenchmarkBackendProcessor/TestProcessTransactionMinimalPayloadValidate-4               8668          6963          -19.67%
BenchmarkBackendProcessor/TestProcessTransactionMinimalPayloadDecode-4                 3499          2564          -26.72%
BenchmarkBackendProcessor/TestProcessTransactionMinimalPayloadTransform-4              7621          5972          -21.64%
BenchmarkBackendProcessor/TestProcessTransactionMinimalPayloadProcessRequest-4         17626         13681         -22.38%
BenchmarkFrontendProcessor/TestProcessTransactionFrontendValidate-4                    16566         13935         -15.88%
BenchmarkFrontendProcessor/TestProcessTransactionFrontendDecode-4                      6901          5595          -18.92%
BenchmarkFrontendProcessor/TestProcessTransactionFrontendTransform-4                   20418         18147         -11.12%
BenchmarkFrontendProcessor/TestProcessTransactionFrontendProcessRequest-4              45221         35125         -22.33%
BenchmarkFrontendProcessor/TestProcessTransactionAugmentedMergeValidate-4              10160         8025          -21.01%
BenchmarkFrontendProcessor/TestProcessTransactionAugmentedMergeDecode-4                3605          3164          -12.23%
BenchmarkFrontendProcessor/TestProcessTransactionAugmentedMergeTransform-4             8231          7139          -13.27%
BenchmarkFrontendProcessor/TestProcessTransactionAugmentedMergeProcessRequest-4        17610         16261         -7.66%
BenchmarkFrontendProcessor/TestProcessTransactionAugmentedValidate-4                   7976          7842          -1.68%
BenchmarkFrontendProcessor/TestProcessTransactionAugmentedDecode-4                     3205          3336          +4.09%
BenchmarkFrontendProcessor/TestProcessTransactionAugmentedTransform-4                  7166          7612          +6.22%
BenchmarkFrontendProcessor/TestProcessTransactionAugmentedProcessRequest-4             16409         15843         -3.45%

benchmark                                                                              old allocs     new allocs     delta
BenchmarkBackendProcessor/TestProcessErrorMinimalServiceValidate-4                     8              8              +0.00%
BenchmarkBackendProcessor/TestProcessErrorMinimalServiceDecode-4                       10             10             +0.00%
BenchmarkBackendProcessor/TestProcessErrorMinimalServiceTransform-4                    38             38             +0.00%
BenchmarkBackendProcessor/TestProcessErrorMinimalServiceProcessRequest-4               56             56             +0.00%
BenchmarkBackendProcessor/TestProcessErrorMinimalProcessValidate-4                     14             14             +0.00%
BenchmarkBackendProcessor/TestProcessErrorMinimalProcessDecode-4                       12             12             +0.00%
BenchmarkBackendProcessor/TestProcessErrorMinimalProcessTransform-4                    41             41             +0.00%
BenchmarkBackendProcessor/TestProcessErrorMinimalProcessProcessRequest-4               67             67             +0.00%
BenchmarkBackendProcessor/TestProcessErrorFullValidate-4                               161            161            +0.00%
BenchmarkBackendProcessor/TestProcessErrorFullDecode-4                                 93             93             +0.00%
BenchmarkBackendProcessor/TestProcessErrorFullTransform-4                              207            207            +0.00%
BenchmarkBackendProcessor/TestProcessErrorFullProcessRequest-4                         461            461            +0.00%
BenchmarkBackendProcessor/TestProcessErrorNullValuesValidate-4                         93             93             +0.00%
BenchmarkBackendProcessor/TestProcessErrorNullValuesDecode-4                           22             22             +0.00%
BenchmarkBackendProcessor/TestProcessErrorNullValuesTransform-4                        87             87             +0.00%
BenchmarkBackendProcessor/TestProcessErrorNullValuesProcessRequest-4                   202            202            +0.00%
BenchmarkBackendProcessor/TestProcessErrorAugmentedIPValidate-4                        8              8              +0.00%
BenchmarkBackendProcessor/TestProcessErrorAugmentedIPDecode-4                          15             15             +0.00%
BenchmarkBackendProcessor/TestProcessErrorAugmentedIPTransform-4                       42             42             +0.00%
BenchmarkBackendProcessor/TestProcessErrorAugmentedIPProcessRequest-4                  65             65             +0.00%
BenchmarkBackendProcessor/TestProcessErrorMinimalPayloadExceptionValidate-4            0              0              +0.00%
BenchmarkBackendProcessor/TestProcessErrorMinimalPayloadExceptionDecode-4              9              9              +0.00%
BenchmarkBackendProcessor/TestProcessErrorMinimalPayloadExceptionTransform-4           36             36             +0.00%
BenchmarkBackendProcessor/TestProcessErrorMinimalPayloadExceptionProcessRequest-4      45             45             +0.00%
BenchmarkBackendProcessor/TestProcessErrorMinimalPayloadLogValidate-4                  8              8              +0.00%
BenchmarkBackendProcessor/TestProcessErrorMinimalPayloadLogDecode-4                    9              9              +0.00%
BenchmarkBackendProcessor/TestProcessErrorMinimalPayloadLogTransform-4                 36             36             +0.00%
BenchmarkBackendProcessor/TestProcessErrorMinimalPayloadLogProcessRequest-4            53             53             +0.00%
BenchmarkFrontendProcessor/TestProcessErrorFrontendValidate-4                          72             72             +0.00%
BenchmarkFrontendProcessor/TestProcessErrorFrontendDecode-4                            57             57             +0.00%
BenchmarkFrontendProcessor/TestProcessErrorFrontendTransform-4                         752            752            +0.00%
BenchmarkFrontendProcessor/TestProcessErrorFrontendProcessRequest-4                    881            881            +0.00%
BenchmarkFrontendProcessor/TestProcessErrorFrontendNoSmapValidate-4                    61             61             +0.00%
BenchmarkFrontendProcessor/TestProcessErrorFrontendNoSmapDecode-4                      44             44             +0.00%
BenchmarkFrontendProcessor/TestProcessErrorFrontendNoSmapTransform-4                   714            714            +0.00%
BenchmarkFrontendProcessor/TestProcessErrorFrontendNoSmapProcessRequest-4              819            819            +0.00%
BenchmarkFrontendProcessor/TestProcessErrorFrontendMinifiedSmapValidate-4              105            105            +0.00%
BenchmarkFrontendProcessor/TestProcessErrorFrontendMinifiedSmapDecode-4                56             56             +0.00%
BenchmarkFrontendProcessor/TestProcessErrorFrontendMinifiedSmapTransform-4             426            426            +0.00%
BenchmarkFrontendProcessor/TestProcessErrorFrontendMinifiedSmapProcessRequest-4        587            587            +0.00%
BenchmarkFrontendProcessor/TestProcessErrorAugmentedUserAgentAndIPValidate-4           8              8              +0.00%
BenchmarkFrontendProcessor/TestProcessErrorAugmentedUserAgentAndIPDecode-4             12             12             +0.00%
BenchmarkFrontendProcessor/TestProcessErrorAugmentedUserAgentAndIPTransform-4          38             38             +0.00%
BenchmarkFrontendProcessor/TestProcessErrorAugmentedUserAgentAndIPProcessRequest-4     58             58             +0.00%
…